requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Werkzeug==2.0.1
Pillow==8.3.1
Flask-Login==0.5.0
Flask-Babel==2.0.0